+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/ui/property/Animatable.kt
@@ -1,5 +1,7 @@
 package com.bumble.appyx.interactions.core.ui.property
 
+import kotlinx.coroutines.CoroutineScope
+
 interface Animatable<T> {
     suspend fun animateTo(
         scope: CoroutineScope,
