+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/transitionmodel/BaseInterpolator.kt
@@ -10,12 +10,24 @@ import com.bumble.appyx.interactions.core.ui.MatchedProps
 import com.bumble.appyx.interactions.core.ui.property.Animatable
 import com.bumble.appyx.interactions.core.ui.property.HasModifier
 import com.bumble.appyx.interactions.core.ui.property.Interpolatable
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.runBlocking
 
 abstract class BaseInterpolator<NavTarget : Any, ModelState, Props>(private val defaultProps: () -> Props) :
     Interpolator<NavTarget, ModelState> where Props : BaseProps, Props : HasModifier, Props : Interpolatable<Props>, Props : Animatable<Props> {
 
     private val cache: MutableMap<String, Props> = mutableMapOf()
+    private val animations: MutableMap<String, Boolean> = mutableMapOf()
+    private val isAnimating: MutableStateFlow<Boolean> = MutableStateFlow(false)
+    override fun isAnimating(): StateFlow<Boolean> =
+        isAnimating
+
+    fun updateAnimationState(key: String, isAnimating: Boolean) {
+        animations[key] = isAnimating
+        this.isAnimating.update { isAnimating || animations.any { it.value } }
+    }
 
     abstract fun ModelState.toProps(): List<MatchedProps<NavTarget, Props>>
 
@@ -29,11 +41,16 @@ abstract class BaseInterpolator<NavTarget : Any, ModelState, Props>(private val 
             FrameModel(
                 navElement = t1.element,
                 modifier = elementProps.modifier.composed {
-                        LaunchedEffect(update) {
-                            elementProps.animateTo(this, t1.props)
-                        }
-                        this
-                    },
+                    LaunchedEffect(update) {
+                        elementProps.animateTo(
+                            scope = this,
+                            props = t1.props,
+                            onStart = { updateAnimationState(t1.element.id, true) },
+                            onFinished = { updateAnimationState(t1.element.id, false) },
+                        )
+                    }
+                    this
+                },
                 progress = 1f,
             )
         }
